Move `ConnectionProxy` to `service` package
