+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/ServiceInstance.kt
@@ -1,6 +1,5 @@
 package net.mullvad.mullvadvpn.service
 
-import net.mullvad.mullvadvpn.dataproxy.ConnectionProxy
 import net.mullvad.talpid.ConnectivityListener
 
 data class ServiceInstance(
